chore(about2): /about2 -> about
