+++ cypress/integration/about.spec.ts
@@ -3,7 +3,7 @@ import { visitWithStatusRetries } from "../helpers/visitWithStatusRetries"
 describe("About", () => {
   it("renders page content", () => {
     visitWithStatusRetries("about")
-    cy.get("h1").should("contain", "For the Love of Art")
+    cy.get("h1").should("contain", "The Future of Art Collecting")
     cy.title().should("eq", "About | Artsy")
   })
 })
